職員削除・払い戻しエラー処理追加

+++ baseball/src/staff/StaffDeleteComplete.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import bean.Staff;
 import dao.StaffDAO;
 
 @WebServlet("/staff/StaffDeleteComplete")
@@ -19,16 +20,24 @@ public class StaffDeleteComplete extends HttpServlet {
     	HttpSession session=request.getSession();
     	String[] staffList=(String[])session.getAttribute("list");
 
-    	StaffDAO DAO=new StaffDAO();
+    	Staff staff=(Staff)session.getAttribute("staff");
+//    	ログインしているか
+    	if(staff == null){
+    		request.setAttribute("sessionOut","1");
+    		request.getRequestDispatcher("login.jsp").forward(request, response);
+    		return;
+    	}else{
+	    	StaffDAO DAO=new StaffDAO();
 
-    	for(String id : staffList){
-    		try {
-				DAO.deleteStaff(id);
-			} catch (Exception e) {
-				//	既に消されている場合はそのまま処理を続ける
-			}
-    	}
-    	session.removeAttribute("list");
-        request.getRequestDispatcher("/staff/staffDeleteComplete.jsp").forward(request, response);
+	    	for(String id : staffList){
+	    		try {
+					DAO.deleteStaff(id);
+				} catch (Exception e) {
+					//	既に消されている場合はそのまま処理を続ける
+				}
+	    	}
+	    	session.removeAttribute("list");
+	        request.getRequestDispatcher("/staff/staffDeleteComplete.jsp").forward(request, response);
+	    }
     }
 }
